Adding archetype templates to docs site
This commit provides a set of templates that users can easily use to create new
pages on our website.

For example, run the following command to add a new Learn page:

~~~
$ hugo new learn/my-new-page.adoc
~~~

And Hugo will use the archetypes/learn.adoc template as a basis.

To add a new pattern, run the following command:

~~~
$ hugo new --kind pattern patterns/my-new-pattern
~~~

And Hugo will use the templates from archetypes/pattern to create a new
pattern in the patterns section.

For more information, see https://gohugo.io/content-management/archetypes/