Updated vault info
The doc was not fully uptodate, relative to the last changes.
Amend that, add a command to show the full json for the vault keys
and surround the storing of the unseal keys sentence in a Note.

Also make sure we use the wrapper when invoking make commands in all
examples

Closes #344