Fix for newer hugo versions
As discovered by Jonny with Hugo versions >= 0.125.0 the website build
breaks with:

    Error: error building site: render: failed to render pages: render
    of "section" failed: "/site/layouts/patterns/list.html:3:9": execute
    of template failed: template: patterns/list.html:3:9: executing
    "main" at <partial "patterns-browser.html" .>: error calling
    partial: "/site/layouts/partials/patterns-browser.html:41:100":
    execute of template failed: template:
    partials/patterns-browser.html:41:100: executing
    "partials/patterns-browser.html" at <urlize>: error calling urlize:
    unable to cast map[string]interface {}{"rh-ocp":interface {}(nil)}
    of type map[string]interface {} to string

Reason for this seem to be the macros used in rh_products in this file
here. Likely because they are defined in
include::modules/comm-attributes.adoc[] but that was never included in
this file.

I tried adding include::modules/comm-attributes.adoc[]  at the very top
of the file (so macros would be defined) but that either breaks the
build or just does not render correctly.